+++ lib/koala/http_service/uploadable_io.rb
@@ -1,4 +1,3 @@
-require "net/http/post/multipart"
 require "tempfile"
 
 module Koala