+++ spec/support/koala_test.rb
@@ -238,11 +238,11 @@ module KoalaTest
         # JRuby doesn't support typhoeus on Travis
         unless defined? JRUBY_VERSION
           require adapter
-          require 'typhoeus/adapters/faraday' if adapter.to_s == "typhoeus"
+          require "faraday/#{adapter}"
           Faraday.default_adapter = adapter.to_sym
         end
-      rescue ParserError
-        puts "Unable to load adapter #{adapter}, using Net::HTTP."
+      rescue => e
+        puts "Unable to load adapter #{adapter}, using Net::HTTP. #{e.class} #{e.message}"
       ensure
         @adapter_activation_attempted = true
       end